requirements.txt
openai
pydantic
Django
djangorestframework
django-filter
django-cors-headers
psycopg2
djangorestframework-simplejwt
djoser
gunicorn
django-storages[google]
google-cloud-storage
wheel 
gunicorn 
redis 
celery
google
python-dotenv